Struct uxi::Bundle

pub struct Bundle<C: Send> { /* private fields */ }

Bundle is a packet containing all the relevant context necessary for a Command invocation. It provides access to the values of the flags provided to the command during invocation, the user specific context, and the inbuilt context for use in a Command’s run function. A given Bundle is tied to a Command’s invocation and can’t be used outside that context.

Implementations

impl<C: Send> Bundle<C>

pub fn lock(&self) -> MutexGuard<'_, BundledCtx<C>>

lock locks the internal mutex of the Bundle and returns a mutex-locked BundledCtx which allows access to the user provided and inbuilt contexts stored in the Client. The mutex can be unlocked by calling the drop function on the variable storing the mutex guard.

// bundle: Bundle<C>
let context = bundle.lock(); // Locking the mutex.
drop(context);               // Unlocking the mutex.

You can lock the mutex again by calling the lock function on the Bundle. Remember to unlock the mutex when not in use in parallel or long-running Commands so that other Commands don’t get stuck trying to access the contexts. In other cases, the mutex is unlocked when the Command ends.

pub fn is_flag_set(&self, name: &str) -> bool

is_flag_set checks if a flag with the given name was provided to the Command during invocation. It works for all types of flags, not just boolean ones.

pub fn get_single_flag(&self, name: &str) -> Option<String>

get_single_flag gets the value provided to a single flag during the Command’s invocation. It returns None if the flag was not set during invocation.

pub fn get_array_flag(&self, name: &str) -> Option<Vec<String>>

get_single_flag gets the value provided to an array or a variadic flag during the Command’s invocation. It returns None if the flag was not set during invocation.